Added a profiles page so the user could track past tests and fixed the create questions UI.

+++ frontend/src/Pages/TestHomePage.tsx
@@ -39,6 +39,12 @@ function Home() {
     }, [])
 
     async function deleteTest(testId: string) {
+        const confirmDelete = window.confirm("Are you sure you want to delete this test? Action cannot be undone.")
+
+        if(!confirmDelete){
+            return;
+        }
+
         const {error} = await supabase
             .from("Test")
             .delete()
